refactor(py): move endpoints to genkit.core.endpoints; prepping for flows endpoint
CHANGELOG:
- [ ] Move the reflection API endpoints implementation to
  genkit.core.endpoints.
