+++ py/packages/genkit/src/genkit/ai/veneer.py
@@ -114,6 +114,7 @@ from genkit.blocks.model import (
 )
 from genkit.blocks.prompt import to_generate_action_options
 from genkit.core.action import ActionKind, ActionRunContext
+from genkit.core.endpoints.reflection import make_reflection_server
 from genkit.core.environment import is_dev_environment
 from genkit.core.schema import to_json_schema
 from genkit.core.typing import (
@@ -123,7 +124,6 @@ from genkit.core.typing import (
     Part,
     ToolChoice,
 )
-from genkit.web.server import make_reflection_server
 
 DEFAULT_REFLECTION_SERVER_SPEC = server.ServerSpec(
     scheme='http', host='127.0.0.1', port=3100
