+++ py/packages/genkit/src/genkit/core/endpoints/reflection.py
